% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureselection.R
\name{featureselection}
\alias{featureselection}
\title{Feature selection}
\usage{
featureselection(x = NULL, y = NULL, method = "iHCT")
}
\arguments{
\item{x}{a matrix of predictor variables}

\item{y}{a vector of binary outcome}

\item{method}{feature selection method, default is iHCT}
}
\value{
\code{featureselection()} returns selected features and other outcomes needed for sample size determination.
}
\description{
This function selects important features from the dataset
}
\examples{

## load data
#pilot.data = readRDS(system.file("extdata", "pilotdata.rds", package = "planningML"))
#x = pilot.data[,-ncol(pilot.data)]
#y = pilot.data$DEPRESSION

## select important features
#features = featureselection(x = x, y = y)
#summary(features)

}
