% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_learningcurve.R
\name{fit_learningcurve}
\alias{fit_learningcurve}
\title{Generate descriptive summary for objects returned by functions in EHRsampling}
\usage{
fit_learningcurve(df, testX, target = NULL)
}
\arguments{
\item{df}{data for learning curve fitting; first column is sample size, second column is AUC measurement.}

\item{testX}{test data for prediction}

\item{target}{target MCC/AUC that you want to achieve}
}
\value{
\code{fit_learningcurve()} returns the estimated power law model for the learning curve.
}
\description{
Generate descriptive summary for objects returned by functions in EHRsampling.
}
