% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.planningML.R
\name{plot.planningML}
\alias{plot.planningML}
\title{Plot sample size dependent AUC or MCC based on number of selected features}
\usage{
\method{plot}{planningML}(x, ...)
}
\arguments{
\item{x}{the output returned by the samplesize function}

\item{...}{ignored arguments}
}
\value{
\code{plot()} returns a scatterplot of sample size dependent performance measurement metrics (AUC or MCC) based on number of selected features
}
\description{
Plot the output returned by samplesize function
}
