% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.planningML.R
\name{summary.planningML}
\alias{summary.planningML}
\title{Generate descriptive summary for objects returned by functions in EHRsampling}
\usage{
\method{summary}{planningML}(object, ...)
}
\arguments{
\item{object}{the object returned by other functions.}

\item{...}{ignored arguments}
}
\value{
\code{summary()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by functions in EHRsampling.
}
