% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_learningcurve.R
\name{fit_learningcurve}
\alias{fit_learningcurve}
\title{Generate descriptive summary for objects returned by functions in EHRsampling}
\usage{
fit_learningcurve(df, testX)
}
\arguments{
\item{df}{data for learning curve fitting; first column is sample size, second column is AUC measurement.}

\item{testX}{test data for prediction and}
}
\value{
\code{summary()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by functions in EHRsampling.
}
